requirements.txt
streamlit
python-dotenv

requests>=2.31.0